requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
alembic==1.13.1

pydantic==2.7.1
email-validator==2.1.1

python-dotenv==1.0.1
httpx==0.27.0

argon2-cffi==23.1.0
